+++ packages/gatsby-plugin-sharp/src/scheduler.js
@@ -1,8 +1,8 @@
 const _ = require(`lodash`)
 const ProgressBar = require(`progress`)
-const existsSync = require(`fs-exists-cached`).sync
+const { existsSync } = require(`fs`)
 const queue = require(`async/queue`)
-const processFile = require(`./process-file`)
+const { processFile } = require(`./process-file`)
 
 const toProcess = {}
 let totalJobs = 0